backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4
openai>=1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
